requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pandas>=1.3.0
scipy>=1.7.0
nibabel>=3.2.0
SimpleITK>=2.1.0
albumentations>=1.1.0
tqdm>=4.62.0